Add PA2 and PF1 metrics and specifications
These metrics are interesting because, in the SRD, the metric measurement
is written such that a specification of the other metric is a
configuration. In the early validate_base framework these are called
"dependent metrics."

Here we've decided to unroll these dependencies by creating a metric for
each metric measurement function configuration. This means there are
dimensions of stretch/design/minimum, and also gri/uzy.

Then the specifications are simply named 'srd' since they are the SRD
specification of that metric.

In cfht_gri.yaml, I've added specifications that inherit from the srd
specifications and mix in the provenance query for CFHT gri validation
data.